fix: migrate categories to event driven references
