--- backend/tevent/src/main/java/com/tbank/tevent/category/dto/CategoryDto.java
@@ -1,11 +1,0 @@
-package com.tbank.tevent.category.dto;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public record CategoryDto(
-        UUID id,
-        String name,
-        UUID eventId,
-        LocalDateTime createdAt
-) {}
