--- backend/tevent/src/main/java/com/tbank/tevent/category/dto/CreateCategoryRequest.java
@@ -1,8 +1,0 @@
-package com.tbank.tevent.category.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-public record CreateCategoryRequest(
-        @NotBlank @Size(max = 255) String name
-) {}
