+++ backend/tevent/src/main/java/com/tbank/tevent/repo/entity/Category.java
@@ -24,10 +24,6 @@ public class Category {
     @Column(nullable = false)
     private String name;
 
-    @Column(name = "user_id", nullable = false)
-    @JdbcTypeCode(SqlTypes.UUID)
-    private UUID userId;
-
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt;
